+++ examples/atomcam2_nerves_app/lib/atomcam2_nerves_app/application.ex
@@ -22,7 +22,8 @@ defmodule Atomcam2NervesApp.Application do
 
         [
           Atomcam2NervesApp.TimeSync,
-          Atomcam2NervesApp.FirmwareHealth
+          Atomcam2NervesApp.FirmwareHealth,
+          Atomcam2NervesApp.NasExporter
         ]
     end
   end